(fix) Improved file structure organization, separated out types into per-component folders
