+++ src/app/components/ForecastModal/types.ts
@@ -1,0 +1,8 @@
+export interface ForecastModalProps {
+  toggleModal: any;
+  isOpen: boolean;
+  weatherHeader: React.DetailedHTMLProps<
+    React.HTMLAttributes<HTMLDivElement>,
+    HTMLDivElement
+  >;
+}
