+++ src/app/components/ZipCode/types.ts
@@ -1,0 +1,4 @@
+export interface LocationProps {
+  id: number;
+  text: string;
+}
