+++ src/app/pages/HomePage/Loadable.tsx
@@ -5,6 +5,6 @@
 import { lazyLoad } from 'utils/loadable';
 
 export const HomePage = lazyLoad(
-  () => import('./index'),
+  () => import('./HomePage'),
   module => module.HomePage,
 );
