--- src/app/pages/HomePage/index.tsx
@@ -1,43 +1,0 @@
-import * as React from 'react';
-import { useSelector } from 'react-redux';
-import { Helmet } from 'react-helmet-async';
-import ZipCodeForm from '../../components/HomePage/ZipCodeForm';
-import ZipCode from '../../components/HomePage/ZipCode';
-
-interface ZipCodeProps {
-  id: number;
-  text: string;
-}
-
-export function HomePage() {
-  const zipCodes = useSelector((state: any) => state.zipCodes);
-
-  const zipCodeList = zipCodes?.map((zipCode: ZipCodeProps) => (
-    <ZipCode key={zipCode.id} id={zipCode.id} text={zipCode.text} />
-  ));
-  return (
-    <>
-      <div className="app">
-        <Helmet>
-          <title>Simple Weather App</title>
-          <meta
-            name="description"
-            content="A simple weather application providing current and future weather conditions"
-          />
-        </Helmet>
-        <h1 className="whiteText">
-          <span role="img" aria-label="crystal ball">
-            🔮
-          </span>
-          &nbsp; Weather Tracker
-        </h1>
-        <div>
-          <ZipCodeForm />
-        </div>
-        <div className="zipCode-list">
-          <ul>{zipCodeList}</ul>
-        </div>
-      </div>
-    </>
-  );
-}
