+++ src/app/pages/HomePage/types.ts
@@ -1,0 +1,4 @@
+export interface ZipCodeProps {
+  id: number;
+  text: string;
+}
